Removed old FC_String class and updated all usages. Now everything uses c-style sprtintf functions
Former-commit-id: fd6dd83046fb54adee44b47a70a782f59cd54260 [formerly d08a1353a8735c77cae21ebceba980126b27e17b]
Former-commit-id: 67b292ad9f7e6f0b03ab83b5c1a1aefba83d06d4

+++ fancy_core/RenderCore_PlatformDX12.cpp
@@ -48,7 +48,7 @@ namespace Fancy {
       case D3D12_COMMAND_LIST_TYPE_COMPUTE: return CommandListType::Compute;
       case D3D12_COMMAND_LIST_TYPE_COPY: return CommandListType::DMA;
       default:
-        ASSERT("Command list type % not supported", aType);
+        ASSERT("Command list type %d not supported", (uint) aType);
         return CommandListType::Graphics;
       }
     }
